biasbox/requirements.txt
Django==5.2.1
django-location-field==2.7.3
